requirements.txt
freegame
python 3.7
